Increase `timeout` Parameter to Accept 64-bit Integers for Larger Timeouts (#239)
* timeout should be a double to support larger values

* cast double to long

* update ios code to use double for timeout

* increment version to 2.3.0

+++ android/src/main/java/com/ocetnik/timer/BackgroundTimerModule.java
@@ -75,7 +75,7 @@ public class BackgroundTimerModule extends ReactContextBaseJavaModule {
     }
 
     @ReactMethod
-    public void setTimeout(final int id, final int timeout) {
+    public void setTimeout(final int id, final double timeout) {
         Handler handler = new Handler();
         handler.postDelayed(new Runnable(){
             @Override
@@ -86,7 +86,7 @@ public class BackgroundTimerModule extends ReactContextBaseJavaModule {
                         .emit("backgroundTimer.timeout", id);
                 }
            }
-        }, timeout);
+        }, (long) timeout);
     }
 
     /*@ReactMethod
